fix(dev-11): address review findings (remove unauthorized primitives)
