+++ src/renderer/src/components/session/OpenInSplitButton.tsx
@@ -1,6 +1,5 @@
 import { useState } from 'react'
 import type { EditorPreset, Session } from '../../../../shared/ipc'
-import { Button, Separator } from '@/components/ui'
 import { useLastOpenInEditor } from '@/hooks/useLastOpenInEditor'
 import {
   DropdownMenu,
@@ -34,6 +33,44 @@ function getEditorLabel(editor: EditorPreset): string {
 function getEditorIcon(): React.ReactNode {
   // Icons can be added here if needed
   return null
+}
+
+interface OutlineButtonProps extends React.ButtonHTMLAttributes<HTMLButtonElement> {
+  size?: 'sm' | 'md' | 'lg'
+  disabled?: boolean
+}
+
+function OutlineButton({
+  children,
+  className = '',
+  size = 'sm',
+  disabled,
+  ...props
+}: OutlineButtonProps): React.JSX.Element {
+  const sizeClasses = {
+    sm: 'h-7 px-2.5 text-[12px]',
+    md: 'h-8 px-3 text-[13px]',
+    lg: 'h-9 px-4 text-[14px]'
+  }
+
+  const buttonClasses = [
+    'inline-flex items-center justify-center gap-2',
+    'font-body font-medium rounded-md',
+    'transition-colors duration-75 outline-none border',
+    'focus-visible:ring-1 focus-visible:ring-accent focus-visible:ring-offset-1 focus-visible:ring-offset-transparent',
+    'data-[disabled]:opacity-50 data-[disabled]:cursor-not-allowed data-[disabled]:pointer-events-none',
+    'border-op-border bg-transparent hover:bg-op-zinc-800 text-op-zinc-200 active:bg-op-zinc-900',
+    sizeClasses[size],
+    className
+  ]
+    .filter(Boolean)
+    .join(' ')
+
+  return (
+    <button className={buttonClasses} disabled={disabled} {...props}>
+      {children}
+    </button>
+  )
 }
 
 export function OpenInSplitButton({
@@ -74,8 +111,7 @@ export function OpenInSplitButton({
   return (
     <div className="flex items-center h-7">
       <div className="flex rounded-md overflow-hidden border border-op-border">
-        <Button
-          variant="outline"
+        <OutlineButton
           size="sm"
           onClick={handlePrimaryClick}
           disabled={disabled}
@@ -84,21 +120,20 @@ export function OpenInSplitButton({
         >
           {icon}
           <span>{editorLabel}</span>
-        </Button>
+        </OutlineButton>
 
-        <Separator orientation="vertical" className="h-5 self-center bg-op-border" />
+        <div className="shrink-0 w-px h-5 self-center bg-op-border" />
 
         <DropdownMenu open={isDropdownOpen} onOpenChange={setIsDropdownOpen}>
           <DropdownMenuTrigger asChild>
-            <Button
-              variant="outline"
+            <OutlineButton
               size="sm"
               disabled={disabled}
               title={disabled ? 'No working directory' : 'Choose editor'}
               className="rounded-l-none border-l-0 px-2"
             >
               <ChevronDown size={14} strokeWidth={1.75} />
-            </Button>
+            </OutlineButton>
           </DropdownMenuTrigger>
 
           <DropdownMenuContent align="end">
