+++ src/renderer/src/components/ui/index.ts
@@ -6,9 +6,6 @@ export { ConfigBadge } from './ConfigBadge'
 export type { ConfigBadgeVariant } from './ConfigBadge'
 export { ToastProvider, useToast } from './Toast'
 export type { ToastOptions, ToastAction } from './Toast'
-export { Button } from './Button'
-export type { ButtonProps, ButtonVariant, ButtonSize } from './Button'
-export { Separator } from './Separator'
 export {
   DropdownMenu,
   DropdownMenuTrigger,
